+++ tests/integration/common.ts
@@ -1,7 +1,7 @@
 import { signInWithEmailAndPassword } from "firebase/auth"
 import { nanoid } from "nanoid"
 import { auth } from "../../components/firebase"
-import { BillContent } from "../../functions/src/bills/types"
+import { Bill, BillContent } from "../../functions/src/bills/types"
 import { testDb, testTimestamp } from "../testUtils"
 
 export async function signInUser(email: string) {
@@ -51,4 +51,9 @@ export async function expectStorageUnauthorized(work: Promise<any>) {
     .catch(e => e)
   expect(e.code).toBe("storage/unauthorized")
   console.warn = warn
+}
+
+export async function getBill(id: string): Promise<Bill> {
+  const doc = await testDb.doc(`/generalCourts/192/bills/${id}`).get()
+  return doc.data() as any
 }
